tests/wpt/tests/html/semantics/forms/constraints/inputwillvalidate.html

<!DOCTYPE html>
<html><head>
    <title>willValidate property on the input element</title>
    <meta content="text/html; charset=UTF-8" http-equiv="content-type">
    <meta content="willValidate property on the input element" name="description">
    <link href="https://html.spec.whatwg.org/multipage/#dom-cva-willvalidate" rel="help">
  </head>
  <body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <div id="log"></div>

    <form action="http://www.example.com/" style="display : none">
      <input required="" type="text">
      <input disabled="" type="text">
    </form>

    <script type="text/javascript">

      test(function() {assert_true(document.getElementsByTagName("input")[0].willValidate)}, "willValidate property returns true when required attribute exists");
      test(function() {assert_false(document.getElementsByTagName("input")[1].willValidate)}, "willValidate property returns false when disabled attribute exists");

    </script>

</body></html>
